Port some more of chrome/ to Linux.
Original review: http://codereview.chromium.org/4247

Patch from Pawel Hajdan Jr.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2793 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/history/visit_tracker.cc
@@ -11,8 +11,8 @@ namespace history {
 // When the list gets longer than 'MaxItems', CleanupTransitionList will resize
 // the list down to 'ResizeTo' size. This is so we only do few block moves of
 // the data rather than constantly shuffle stuff around in the vector.
-static const int kMaxItemsInTransitionList = 96;
-static const int kResizeBigTransitionListTo = 64;
+static const size_t kMaxItemsInTransitionList = 96;
+static const size_t kResizeBigTransitionListTo = 64;
 COMPILE_ASSERT(kResizeBigTransitionListTo < kMaxItemsInTransitionList,
                max_items_must_be_larger_than_resize_to);
