+++ chrome/browser/history/visitsegment_database.cc
@@ -240,7 +240,7 @@ void VisitSegmentDatabase::QuerySegmentUsage(
   // does as well.
 
   // How many results we return, as promised in the header file.
-  const int kResultCount = 9;
+  const size_t kResultCount = 9;
 
   // Gather all the segment scores:
   SQLITE_UNIQUE_STATEMENT(statement, GetStatementCache(),
